+++ CHANGES.md
@@ -574,6 +574,8 @@ v0.13.0 (2018-04-27)
   subsets to define the areas in which to compute profiles rather
   than custom ROIs. [#1635]
 
+* Remove ``StandaloneImageViewer`` class. [#1663]
+
 * Added support for PySide2 and remove support for PyQt4 and
   PySide. [#1662]